Move blacklist filtering function into model_utils
